# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Customer:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
    type: object
  entity.Employee:
    properties:
      address:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
    type: object
  entity.Login:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: sdesain25@gmail.com
    name: API Support
    url: https://uqonstd.xyz/
  description: A documentation how to access the api's routes on your laundry application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Laundry API Documentation
  version: "1.0"
paths:
  /customers:
    get:
      consumes:
      - application/json
      description: Retrieve a list of customers
      parameters:
      - description: Filter by customer name
        in: query
        name: name
        type: string
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: Get Customers
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Create a new customer
      parameters:
      - description: Customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/entity.Customer'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create Customer
      tags:
      - Customers
  /customers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: customer successfully deleted
          schema:
            type: string
        "400":
          description: error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete Customer
      tags:
      - Customers
    get:
      consumes:
      - application/json
      description: Retrieve a customer by ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Customer by ID
      tags:
      - Customers
    put:
      consumes:
      - application/json
      description: Update an existing customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/entity.Customer'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update Customer
      tags:
      - Customers
  /employees:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all employees
      parameters:
      - description: Filter by employee name
        in: query
        name: name
        type: string
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Employee data
          schema:
            $ref: '#/definitions/entity.Employee'
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: List all employees
      tags:
      - Employees
    post:
      consumes:
      - application/json
      description: Add a new employee to the database
      parameters:
      - description: Employee data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/entity.Employee'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Employee data
          schema:
            items:
              $ref: '#/definitions/entity.Employee'
            type: array
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: Create a new employee
      tags:
      - Employees
  /employees/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an employee from the database
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Employee data
          schema:
            $ref: '#/definitions/entity.Employee'
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: Delete an employee
      tags:
      - Employees
    get:
      description: Retrieve employee data by ID
      operationId: id
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Employee data
          schema:
            $ref: '#/definitions/entity.Employee'
        "400":
          description: Bad Request
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Get employee by ID
      tags:
      - Employees
    put:
      consumes:
      - application/json
      description: Update the details of an existing employee in the database
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employee data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/entity.Employee'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Employee data
          schema:
            $ref: '#/definitions/entity.Employee'
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: Update an existing employee
      tags:
      - Employees
  /login:
    post:
      consumes:
      - application/json
      description: 'Authenticate an employee and generate a token exp:. email : pegi@gmail.com,
        password: @Emka123'
      parameters:
      - description: Login data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/entity.Login'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            $ref: '#/definitions/entity.Employee'
        "400":
          description: Bad Request
      summary: Login an employee
      tags:
      - Auth Token
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"